@@ -97,15 +97,16 @@ func (t *quicListener) serve(ctx context.Context) error {
 	}
 	defer func() { _ = udpConn.Close() }()
 
-	svc, conn := stun.New(t.cfg, t, udpConn)
-	defer conn.Close()
+	quicTransport := &quic.Transport{Conn: udpConn}
 
-	go svc.Serve(ctx)
+	// svc, conn := stun.New(t.cfg, t, udpConn)
+	// defer conn.Close()
+	// go svc.Serve(ctx)
 
-	t.registry.Register(t.uri.Scheme, conn)
-	defer t.registry.Unregister(t.uri.Scheme, conn)
+	t.registry.Register(t.uri.Scheme, quicTransport)
+	defer t.registry.Unregister(t.uri.Scheme, quicTransport)
 
-	listener, err := quic.Listen(conn, t.tlsCfg, quicConfig)
+	listener, err := quicTransport.Listen(t.tlsCfg, quicConfig)
 	if err != nil {
 		l.Infoln("Listen (BEP/quic):", err)
 		return err
